sc-1273-refactor publishing drawer and form, update styles to match design
